Store index numbers as integers
Given that a row will not likely be longer than 15 elements
using integers can easily store row indexes

+++ src/main/java/com/mobiquityinc/packer/Packer.java
@@ -41,7 +41,7 @@ public class Packer {
         StringBuilder resultBuilder = new StringBuilder();
         specifications.forEach(spec -> {
 
-            List<Long> indexNumbers = packagingResolver.resolvePackaging(spec);
+            List<Integer> indexNumbers = packagingResolver.resolvePackaging(spec);
 
             if (indexNumbers.isEmpty()) {
                 resultBuilder.append("-\n");
